requirements.txt
flask
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
python-dotenv
gunicorn
python-dateutil